+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-# 1.7.0 / 2014-03-25
-* [FEATURE] Sub out Friendship for UserFollow
+# 2.0.0 / 2014-03-25
+* [FEATURE] Sub in UserFollow and UserConnection for Friendship
+* [FEATURE] Sub in FollowerBadge for BefrienderBadge
 
 # 1.6.16 / 2014-03-24
 * [FEATURE] Add slug to collections